requirements.txt
asgiref==3.5.2
autopep8==2.0.1
Django==4.1.4
django-crispy-forms==1.14.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1